+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 -------------------
 
 - `classes.new_instance` method can take an options list now as well
+- `classes.get_enum` method returns the instance of an enum item
+- `typeconv.string_list_to_python` method converts a `java.util.List` containing strings into a Python list
 
 
 0.1.16 (2020-12-26)